refactor(hush-cli): extract install from pkg_cli
Move the pkg install cluster (cmd_pkg_install router, local + registry
handlers, rollback machinery, version selection, fingerprint recompute,
size-limited download reader) into pkg_cli/install.rs. cmd_pkg_install
and recompute_installed_content_fingerprint widened to pub(super) for
the dispatcher / verify handler; read_archive_identity,
requested_identity_matches_install, create_install_rollback_backup,
restore_install_from_backup, select_default_registry_version and
InstallRollbackBackup (+ backup_path field) widened to pub(super) for
tests. mod.rs imports trimmed; archive + is_file_source are now
test-only there.
